Remove Heretic BLOCKMAP limit (#443)
* Remove Heretic BLOCKMAP limit

* Change longs to int32_t for blockmap limit removal

Make sure that  blockmap limit removal code is in line with crispy-doom

* Add commented out assignment of count back to P_LoadBlockMap

+++ src/heretic/p_sight.c
@@ -91,7 +91,7 @@ boolean PTR_SightTraverse(intercept_t * in)
 boolean P_SightBlockLinesIterator(int x, int y)
 {
     int offset;
-    short *list;
+    int32_t *list;
     line_t *ld;
     int s1, s2;
     divline_t dl;
